First commit with JPA, Guice as DAL.
Never got this to work properly. E.g. if a recipe with some Part objects is created will a get request for these part object only return the correct information if it is the same thread that created the parts objects that handling the get request.

1:  curl -v -H "Content-Type: application/json" -X POST -d {"title":"test1","comment":"comment1"} http://localhost:8080/recipes
2:  curl -v -H "Content-Type: application/json" -X POST -d {"name":"part 1 to recipy 1","instructions": ["instruction1", "instruction2"]} http://localhost:8080/recipes/1/parts
3:  curl -v -H "Content-Type: application/json" -X POST -d {"name":"part 2 to recipy 1","instructions": ["instructionzz1", "instructionzz2"]} http://localhost:8080/recipes/1/parts
4:  curl -v -H "Accept: application/json" http://localhost:8080/recipes/1/parts

The "4" will give different results.